Conflicts:
	.idea/workspace.xml

+++ sqliteinterupt/controldeamon.py
@@ -6,6 +6,7 @@ import os
 import threading
 import inputdeamon
 from inputdeamon.inputDeamon import inputmainlink
+from outputdeamon.outputdeamon import outputmainlink
 
 # class DeamonThreads(threading.Thread):
 #      def __init__(self,inptype):
@@ -16,10 +17,20 @@ from inputdeamon.inputDeamon import inputmainlink
 #             inputmainlink()
 
 
+def initiateprocess():
+    controlpid=currentDb.getcontrolpid()
+    inputpid=currentDb.getinputpid()
+    outputpid=currentDb.getoutputpid()
+    os.kill(controlpid,9)
+    os.kill(inputpid, 9)
+    os.kill(outputpid, 9)
 
+
+initiateprocess()
 currentDb=initdbclass()
 print "----*----"
 #inputthread=None
+currentDb.updatecontrolpid(os.getgid())
 while(1):
     print "loop terminated"
 
@@ -27,29 +38,34 @@ while(1):
     if newinputpid == 0:
         inputmainlink()
     else:
-        #print('Hello from parent', os.getpid(), newinputpid)
-        while(1):##For get the data from control
-            #inputthread=DeamonThreads(1)
-            isclient=True
-            datarow=currentDb.getnodedata("cntrldeamon")
-            controlip=datarow[2]
-            controlport=datarow[3]
-            type=datarow[4]
-            if(type=="client"):
-                controllink=DeamonLinkFactory("client").getConnection(controlip,controlport)
-                while(not controllink.connect()):
-                    time.sleep(2)
-                    print "Control Link Connection error occured"
-                try:
-                    #inputthread.start()
-                    controllink.getdata()
-                    print "get data returened"
-                    time.sleep(1)
-                    os.kill(newinputpid)
-                    break;
-                except Exception,e:
-                    #print "Control Server Link gone ",e
-                    controllink.disconnect()
+        newoutputpid = os.fork()
+        if newoutputpid == 0:
+            outputmainlink()
+        else:
+            #print('Hello from parent', os.getpid(), newinputpid)
+            while(1):##For get the data from control
+                #inputthread=DeamonThreads(1)
+                isclient=True
+                datarow=currentDb.getnodedata("cntrldeamon")
+                controlip=datarow[2]
+                controlport=datarow[3]
+                type=datarow[4]
+                if(type=="client"):
+                    controllink=DeamonLinkFactory("client").getConnection(controlip,controlport)
+                    while(not controllink.connect()):
+                        time.sleep(2)
+                        print "Control Link Connection error occured"
+                    try:
+                        #inputthread.start()
+                        controllink.getdata()
+                        print "get data returened"
+                        time.sleep(1)
+                        os.kill(newinputpid)
+                        os.kill(newoutputpid)
+                        break;
+                    except Exception,e:
+                        #print "Control Server Link gone ",e
+                        controllink.disconnect()
 
 def child():
     for i in range(0,50):
@@ -70,4 +86,5 @@ def parent():
                 #print('Hello from child', os.getpid())
                 print('Hello from parent', os.getpid(), newpid)
         if input() == 'q': break
+
